Fix state machine steps and increase merkkijono
Because program uses the button interruption to change the state
machine state to get back to WAITING_HOME state whatever the
state is, this causes issue when task execution continues after
interruption. Task writes the programState in the end so it will
overwrite the state set previously in interruption.

Merkkijono char array was too small for the accelometer data and
most definitely caused some malfunctioning.